refactor: create the `alloc` module
